[spaceship] improve client (mainly: login) (#13736)
* Add initial information if 2FA or 2step is required

* extract two step handling into own method

* move select_device method

* improve 2-step-validation and 2FA output

* better error message in case spaceship can't handle response from Apple

* adjust code to match with similar code in handle_two_factor

* rename two_step_client to two_step_or_fator_client as we don't know yet if 2step or 2fa - it handles both

* move methods around and add more comment groups

* comments on how we get from `login` to `send_shared_login_request` via the helper method `do_login` and sub classes

* additional verbose output if credentials are loaded from keychain

* additional output when session is not valid any more

* rubocop
