+++ spaceship/lib/spaceship/two_step_or_factor_client.rb
@@ -1,14 +1,18 @@
-require 'tempfile'
-
 require_relative 'globals'
 require_relative 'tunes/tunes_client'
 require_relative 'tunes/recovery_device'
 
 module Spaceship
   class Client
-    def handle_two_step(response)
+    def handle_two_step_or_factor(response)
       @x_apple_id_session_id = response["x-apple-id-session-id"]
       @scnt = response["scnt"]
+
+      puts("")
+      puts("Two-step Verification (4 digits code) or Two-factor Authentication (6 digits code) is enabled for account '#{self.user}'")
+      puts("More information about Two-step Verification (4 digits code): https://support.apple.com/en-us/HT204152")
+      puts("More information about Two-factor Authentication (6 digits code): https://support.apple.com/en-us/HT204915")
+      puts("")
 
       r = request(:get) do |req|
         req.url("https://idmsa.apple.com/appleauth/auth")
@@ -16,105 +20,41 @@ module Spaceship
       end
 
       if r.body.kind_of?(Hash) && r.body["trustedDevices"].kind_of?(Array)
-        if r.body.fetch("securityCode", {})["tooManyCodesLock"].to_s.length > 0
-          raise Tunes::Error.new, "Too many verification codes have been sent. Enter the last code you received, use one of your devices, or try again later."
-        end
-
-        old_client = (begin
-                        Tunes::RecoveryDevice.client
-                      rescue
-                        nil # since client might be nil, which raises an exception
-                      end)
-        Tunes::RecoveryDevice.client = self # temporary set it as it's required by the factory method
-        devices = r.body["trustedDevices"].collect do |current|
-          Tunes::RecoveryDevice.factory(current)
-        end
-        Tunes::RecoveryDevice.client = old_client
-
-        puts("Two Step Verification for account '#{self.user}' is enabled")
-        puts("Please select a device to verify your identity")
-        available = devices.collect do |c|
-          "#{c.name}\t#{c.model_name || 'SMS'}\t(#{c.device_id})"
-        end
-        result = choose(*available)
-        device_id = result.match(/.*\t.*\t\((.*)\)/)[1]
-        select_device(r, device_id)
+        handle_two_step(r)
       elsif r.body.kind_of?(Hash) && r.body["trustedPhoneNumbers"].kind_of?(Array) && r.body["trustedPhoneNumbers"].first.kind_of?(Hash)
         handle_two_factor(r)
       else
-        raise "Invalid 2 step response #{r.body}"
+        raise "Although response from Apple indicated activated Two-step Verification or Two-factor Authentication, spaceship didn't know how to handle this response: #{r.body}"
       end
     end
 
-    def handle_two_factor(response)
-      two_factor_url = "https://github.com/fastlane/fastlane/tree/master/spaceship#2-step-verification"
-      puts("Two Factor Authentication for account '#{self.user}' is enabled")
-
-      puts("If you're running this in a non-interactive session (e.g. server or CI)")
-      puts("check out #{two_factor_url}")
-
-      security_code = response.body["securityCode"]
-      # {"length"=>6,
-      #  "tooManyCodesSent"=>false,
-      #  "tooManyCodesValidated"=>false,
-      #  "securityCodeLocked"=>false}
-      code_length = security_code["length"]
-      code = ask("Please enter the #{code_length} digit code: ")
-      puts("Requesting session...")
-
-      # Send securityCode back to server to get a valid session
-      r = request(:post) do |req|
-        req.url("https://idmsa.apple.com/appleauth/auth/verify/trusteddevice/securitycode")
-        req.headers['Content-Type'] = 'application/json'
-        req.body = { "securityCode" => { "code" => code.to_s } }.to_json
-
-        update_request_headers(req)
+    def handle_two_step(r)
+      if r.body.fetch("securityCode", {})["tooManyCodesLock"].to_s.length > 0
+        raise Tunes::Error.new, "Too many verification codes have been sent. Enter the last code you received, use one of your devices, or try again later."
       end
 
-      # we use `Spaceship::TunesClient.new.handle_itc_response`
-      # since this might be from the Dev Portal, but for 2 step
-      Spaceship::TunesClient.new.handle_itc_response(r.body)
-
-      store_session
-
-      return true
-    end
-
-    # Only needed for 2 step
-    def load_session_from_file
-      if File.exist?(persistent_cookie_path)
-        puts("Loading session from '#{persistent_cookie_path}'") if Spaceship::Globals.verbose?
-        @cookie.load(persistent_cookie_path)
-        return true
+      old_client = (begin
+                      Tunes::RecoveryDevice.client
+                    rescue
+                      nil # since client might be nil, which raises an exception
+                    end)
+      Tunes::RecoveryDevice.client = self # temporary set it as it's required by the factory method
+      devices = r.body["trustedDevices"].collect do |current|
+        Tunes::RecoveryDevice.factory(current)
       end
-      return false
-    end
+      Tunes::RecoveryDevice.client = old_client
 
-    def load_session_from_env
-      return if self.class.spaceship_session_env.to_s.length == 0
-      puts("Loading session from environment variable") if Spaceship::Globals.verbose?
-
-      file = Tempfile.new('cookie.yml')
-      file.write(self.class.spaceship_session_env.gsub("\\n", "\n"))
-      file.close
-
-      begin
-        @cookie.load(file.path)
-      rescue => ex
-        puts("Error loading session from environment")
-        puts("Make sure to pass the session in a valid format")
-        raise ex
-      ensure
-        file.unlink
+      puts("Two-step Verification (4 digits code) is enabled for account '#{self.user}'")
+      puts("Please select a device to verify your identity")
+      available = devices.collect do |c|
+        "#{c.name}\t#{c.model_name || 'SMS'}\t(#{c.device_id})"
       end
+      result = choose(*available)
+      device_id = result.match(/.*\t.*\t\((.*)\)/)[1]
+      select_device(r, device_id)
     end
 
-    # Fetch the session cookie from the environment
-    # (if exists)
-    def self.spaceship_session_env
-      ENV["FASTLANE_SESSION"] || ENV["SPACESHIP_SESSION"]
-    end
-
+    # this is extracted into its own method so it can be called multiple times (see end)
     def select_device(r, device_id)
       # Request Token
       r = request(:put) do |req|
@@ -133,9 +73,8 @@ module Spaceship
       # Send token back to server to get a valid session
       r = request(:post) do |req|
         req.url("https://idmsa.apple.com/appleauth/auth/verify/device/#{device_id}/securitycode")
-        req.body = { "code" => code.to_s }.to_json
         req.headers['Content-Type'] = 'application/json'
-
+        req.body = { "code" => code.to_s }.to_json
         update_request_headers(req)
       end
 
@@ -167,6 +106,41 @@ module Spaceship
 
         raise ex
       end
+
+      store_session
+
+      return true
+    end
+
+    def handle_two_factor(response)
+      two_factor_url = "https://github.com/fastlane/fastlane/tree/master/spaceship#2-step-verification"
+      puts("Two-factor Authentication (6 digits code) is enabled for account '#{self.user}'")
+
+      puts("If you're running this in a non-interactive session (e.g. server or CI)")
+      puts("check out #{two_factor_url}")
+
+      security_code = response.body["securityCode"]
+      # securityCode =
+      # {"length"=>6,
+      #  "tooManyCodesSent"=>false,
+      #  "tooManyCodesValidated"=>false,
+      #  "securityCodeLocked"=>false}
+      code_length = security_code["length"]
+      code = ask("Please enter the #{code_length} digit code: ")
+      puts("Requesting session...")
+
+      # Send securityCode back to server to get a valid session
+      r = request(:post) do |req|
+        req.url("https://idmsa.apple.com/appleauth/auth/verify/trusteddevice/securitycode")
+        req.headers['Content-Type'] = 'application/json'
+        req.body = { "securityCode" => { "code" => code.to_s } }.to_json
+
+        update_request_headers(req)
+      end
+
+      # we use `Spaceship::TunesClient.new.handle_itc_response`
+      # since this might be from the Dev Portal, but for 2 step
+      Spaceship::TunesClient.new.handle_itc_response(r.body)
 
       store_session
